refactor: extract YT logic to hook
